+++ project-base/src/Shopsys/ShopBundle/Model/Administrator/AdministratorData.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Administrator;
+
+use Shopsys\FrameworkBundle\Model\Administrator\AdministratorData as BaseAdministratorData;
+
+class AdministratorData extends BaseAdministratorData
+{
+}